requirements.txt
torch==1.5.1 
numpy==1.18.0 
imutils==0.5.3 
opencv_python==4.2.0.34 
facenet_pytorch==2.3.1 
torchvision==0.6.1 
Pillow==7.2.0 
tqdm==4.49.0
